+++ packages/form-render/src/form-render-core/src/useForm.js
@@ -10,7 +10,6 @@ import {
   clone,
   schemaContainsExpression,
   parseAllExpression,
-  isEmpty,
 } from './utils';
 
 const useForm = props => {
@@ -43,6 +42,7 @@ const useForm = props => {
     isEditing: false, // 是否在编辑状态。主要用于优化体验，用户编辑时减少不必要的运算
     allTouched: false, // 是否所有表单元素都被碰过了（一键开关，用于提交的时候，默认所有都被touch了）
     touchedKeys: [], // 碰过的key（用于submit之前，判断哪些被碰过了）
+    flatten: {}, // schema 的转换结构，便于处理
     finalFlatten: {}, // 表达式等被处理过的flatten，用于渲染
     firstMount: true,
   });
@@ -71,6 +71,7 @@ const useForm = props => {
     isEditing,
     allTouched,
     touchedKeys,
+    flatten,
     finalFlatten,
     firstMount,
     // statusTree, // 和formData一个结构，但是每个元素是 { $touched } 存放那些在schema里无需表达的状态, 看看是否只有touched。目前statusTree没有被使用
@@ -79,6 +80,7 @@ const useForm = props => {
   _errorFields.current = errorFields;
   _outErrorFields.current = outErrorFields;
   _touchedKeys.current = touchedKeys;
+  _flatten.current = flatten;
   _finalFlatten.current = finalFlatten;
 
   const dataFromOutside = props && props.hasOwnProperty('formData');
@@ -108,41 +110,39 @@ const useForm = props => {
   ]);
 
   useEffect(() => {
-    if (schemaRef.current) {
-      // 处理使用setSchemaByPath,使用的是旧flatten 页面不触发更新。
-      let newFlatten = clone(
-        isEmpty(_finalFlatten.current)
-          ? _flatten.current
-          : _finalFlatten.current
-      );
-      if (firstMount) {
-        _flatten.current = flattenSchema(schemaRef.current);
-        setState({ firstMount: false });
-      } else {
-        // 统一的处理expression
-        Object.entries(_flatten.current).forEach(([path, info]) => {
-          if (schemaContainsExpression(info.schema)) {
-            const arrayLikeIndex = path.indexOf(']');
-            const isArrayItem =
-              arrayLikeIndex > -1 && arrayLikeIndex < path.length - 1;
-            const hasRootValue =
-              JSON.stringify(info.schema).indexOf('rootValue') > -1;
-            if (isArrayItem && hasRootValue) {
-              // do nothing
-            } else {
-              newFlatten[path].schema = parseAllExpression(
-                info.schema,
-                _data.current,
-                path
-              );
-            }
-          }
-        });
-      }
-      setState({ finalFlatten: newFlatten });
+    if (schemaRef.current && firstMount) {
+      const flatten = flattenSchema(schemaRef.current);
+      setState({ flatten, firstMount: false });
     }
+  }, [JSON.stringify(schemaRef.current), firstMount]);
+
+  // 统一的处理expression
+  useEffect(() => {
+    if (firstMount) {
+      return;
+    }
+    let newFlatten = clone(_flatten.current);
+    Object.entries(_flatten.current).forEach(([path, info]) => {
+      if (schemaContainsExpression(info.schema)) {
+        const arrayLikeIndex = path.indexOf(']');
+        const isArrayItem =
+          arrayLikeIndex > -1 && arrayLikeIndex < path.length - 1;
+        const hasRootValue =
+          JSON.stringify(info.schema).indexOf('rootValue') > -1;
+        if (isArrayItem && hasRootValue) {
+          // do nothing
+        } else {
+          newFlatten[path].schema = parseAllExpression(
+            info.schema,
+            _data.current,
+            path
+          );
+        }
+      }
+    });
+    setState({ finalFlatten: newFlatten });
   }, [
-    JSON.stringify(schemaRef.current),
+    JSON.stringify(_flatten.current),
     JSON.stringify(_data.current),
     firstMount,
   ]);
@@ -229,10 +229,10 @@ const useForm = props => {
   };
 
   const setSchema = settings => {
-    const newFlatten = clone(_finalFlatten.current);
+    const newFlatten = clone(_flatten.current);
     try {
       Object.keys(settings).forEach(path => {
-        if (!_finalFlatten.current[path]) {
+        if (!_flatten.current[path]) {
           console.error(`path：'${path}' 不存在(form.setSchemaByPath)`);
         } else {
           const newSchema = settings[path];
@@ -246,19 +246,19 @@ const useForm = props => {
           };
         }
       });
-      setState({ finalFlatten: newFlatten });
-      _finalFlatten.current = newFlatten;
+      setState({ flatten: newFlatten });
+      _flatten.current = newFlatten;
     } catch (error) {
       console.error(error, 'setSchema');
     }
   };
 
   const setSchemaByPath = (path, newSchema) => {
-    if (!_finalFlatten.current[path]) {
+    if (!_flatten.current[path]) {
       console.error(`path：'${path}' 不存在(form.setSchemaByPath)`);
       return;
     }
-    const newFlatten = clone(_finalFlatten.current);
+    const newFlatten = clone(_flatten.current);
 
     try {
       const _newSchema =
@@ -266,8 +266,8 @@ const useForm = props => {
           ? newSchema(newFlatten[path].schema)
           : newSchema;
       newFlatten[path].schema = { ...newFlatten[path].schema, ..._newSchema };
-      setState({ finalFlatten: newFlatten });
-      _finalFlatten.current = newFlatten;
+      setState({ flatten: newFlatten });
+      _flatten.current = newFlatten;
     } catch (error) {
       console.error(error, 'setSchemaByPath');
     }
@@ -275,7 +275,7 @@ const useForm = props => {
 
   const getSchemaByPath = path => {
     try {
-      return _finalFlatten.current[path].schema;
+      return _flatten.current[path].schema;
     } catch (error) {
       console.log(error, 'getSchemaByPath');
       return {};
@@ -300,7 +300,11 @@ const useForm = props => {
     let newError = _errorFields.current.filter(item => {
       return item.name.indexOf(path) === -1;
     });
-    setState({ outErrorFields: newError });
+
+    let newOutError = _outErrorFields.current.filter(item => {
+      return item.name.indexOf(path) === -1;
+    });
+    setState({ errorFields: newError, outErrorFields: newOutError });
   };
 
   const getValues = () => {
@@ -312,7 +316,7 @@ const useForm = props => {
   };
 
   const setValues = newFormData => {
-    const newData = transformDataWithBind2(newFormData, _finalFlatten.current);
+    const newData = transformDataWithBind2(newFormData, _flatten.current);
     _setData(newData);
   };